+++ src/hooks/browser/index.ts
@@ -1,5 +1,9 @@
 import useEyeDropper from './useEyeDropper'
+import useUserInfo from './useUserInfo'
+import useDevice from './useDevice'
 
 export default {
   useEyeDropper,
+  useUserInfo,
+  useDevice,
 }