+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ---
 
-- feat: add useEyeDropper & useBluetooth & useUserInfo & usePlatform [#12](https://github.com/janlin002/react-use-ez/issues/12)
+- feat: add useEyeDropper & useBluetooth & useUserInfo & useDevice [#12](https://github.com/janlin002/react-use-ez/issues/12)
 - refactor: modify README
 - refactor: modify folder structure
 - chore: packages add nvmrc, to fixed node version